Finalize Link.Manifest implementation
