+++ opentracing-specialagent/src/main/java/io/opentracing/contrib/specialagent/LibraryFingerprint.java
@@ -23,10 +23,11 @@ import java.io.ObjectOutputStream;
 import java.net.URL;
 import java.net.URLClassLoader;
 import java.util.Arrays;
+import java.util.Objects;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
-import io.opentracing.contrib.specialagent.ReferralScanner.Manifest;
+import io.opentracing.contrib.specialagent.Link.Manifest;
 
 /**
  * A {@link Fingerprint} that represents the fingerprint of a library.
@@ -57,7 +58,7 @@ class LibraryFingerprint extends Fingerprint {
   }
 
   private final ClassFingerprint[] classes;
-  private final Manifest referrals;
+  private final Manifest manifest;
 
   /**
    * Creates a new {@code LibraryFingerprint} with the specified {@code URL}
@@ -65,16 +66,21 @@ class LibraryFingerprint extends Fingerprint {
    *
    * @param parent The parent {@code ClassLoader} to use for resolution of
    *          classes that should not be part of the fingerprint.
+   * @param manifest The {@link Link.Manifest} specifying which classes, methods
+   *          and fields are to be fingerprinted.
    * @param urls The {@code URL} objects referencing JAR files.
+   * @throws NullPointerException If {@code manifest} or {@code urls} is null.
+   * @throws IllegalArgumentException If the number of members in {@code urls}
+   *           is zero.
    * @throws IOException If an I/O error has occurred.
    */
-  LibraryFingerprint(final ClassLoader parent, final Manifest referrals, final URL ... urls) throws IOException {
+  LibraryFingerprint(final ClassLoader parent, final Manifest manifest, final URL ... urls) throws IOException {
     if (urls.length == 0)
       throw new IllegalArgumentException("Number of arguments must be greater than 0");
 
-    this.referrals = referrals;
+    this.manifest = Objects.requireNonNull(manifest);
     try (final URLClassLoader classLoader = new URLClassLoader(urls, parent)) {
-      this.classes = new Fingerprinter(referrals).fingerprint(classLoader);
+      this.classes = new Fingerprinter(manifest).fingerprint(classLoader);
     }
   }
 
@@ -103,6 +109,15 @@ class LibraryFingerprint extends Fingerprint {
   }
 
   /**
+   * Returns the {@code Manifest} of this {@code LibraryFingerprint}.
+   *
+   * @return The {@code Manifest} of this {@code LibraryFingerprint}.
+   */
+  public Manifest getManifest() {
+    return this.manifest;
+  }
+
+  /**
    * Tests whether the runtime represented by the specified {@code ClassLoader}
    * is compatible with this fingerprint.
    *
@@ -113,7 +128,7 @@ class LibraryFingerprint extends Fingerprint {
    *         the runtime is compatible with this fingerprint,
    */
   public FingerprintError[] isCompatible(final ClassLoader classLoader) {
-    return isCompatible(classLoader, new Fingerprinter(referrals), 0, 0);
+    return isCompatible(classLoader, new Fingerprinter(manifest), 0, 0);
   }
 
   /**
