+++ opentracing-specialagent/src/test/java/io/opentracing/contrib/specialagent/LinkManifestTest.java
@@ -1,0 +1,39 @@
+/* Copyright 2019 The OpenTracing Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.opentracing.contrib.specialagent;
+
+import java.io.IOException;
+
+import org.junit.Test;
+import org.objectweb.asm.ClassReader;
+import org.objectweb.asm.Type;
+
+@SuppressWarnings("unused")
+public class LinkManifestTest {
+  private Type memberField = Type.BOOLEAN_TYPE;
+  private Type memberMethod = Type.getType(LinkManifestTest.class);
+
+  static {
+    final int expandFrames = ClassReader.EXPAND_FRAMES;
+    if (expandFrames == -1)
+      throw new ExceptionInInitializerError();
+  }
+
+  @Test
+  public void test() throws IOException {
+    new Link.Manifest().include(ClassLoader.getSystemClassLoader(), getClass().getName().replace('.', '/').concat(".class"));
+  }
+}
